Permit the output window to be shown inside the rift on linux. Add information about running it on linyx inside the README file
The linux compatibility is a WIP and is clunky and 100% unfonctional. It makes stanges call to system to mach the keyboard layout used under windows (usefull for non-qwerty layout users), hardcode the window behaviour and so on.
You'll have to get the rift to be the 2nd screen, non overlapping with the main one, and in the case of the DK2, to be rotated correctly (left seems to works)
Everything is tested on a Nvidia GTX 980 powered workstation with the rift plugged on the DVI-I-1 port. (This is a non-reference card.)