correctly initialize OVR and prevent crash
The engine is still unable to render for the rift at this stage.
I'll have to modify some code inside of OGRE for it to run...
This is not optimal, at all.

I may need to staticly link a modified RenderSystem_GL to be able to
Set the swap textures created by the Oculus SDK. At least I'm getting
texture address so it's on it's way...