SaveData Object can report changed state
AnnSaveData set a "changed" boolean to true when you set a value to a key.
This boolean is reset to false by AnnFileWriter::write() and
AnnFileReader::read()

The method AnnSaveData::hasUnsavedChages() returns true if the content of
the data object isn't saved on disc. If object is destroyed (ad program
terminate), changes will be lost.