correct texture copy bug with on screen console opened
the Ogre::Texture::copyToTexture method used before seems to have a
weird bug that it affects other texture.

The problem is circumvented by doing the copy via direct memory mapping of
the underlying buffers
There's no actuall check done on buffer sizes. But since the background is manually loaded, the size of the 2 textures are identical. If they were not, some scalling should be done,
This would be problematic because would require actually talking to the ram, this being really slow