make the events buffered and dispateched at event update call
This makes the eventManager become owner of the event object when submitting it for dispatch. Event will be sent at next event manager update.

The only bothering thing is that I'm using a shared_ptr to send the event. This is fine functionality wise, but not really in term of syntax, making the use of "make_shared" mandatory